[CLN] microsoft_calendar: split test file and clean up code
