+++ addons/google_calendar/tests/__init__.py
@@ -4,4 +4,5 @@ from . import test_google_event
 from . import test_sync_common
 from . import test_sync_google2odoo
 from . import test_sync_odoo2google
+from . import test_sync_odoo2google_mail
 from . import test_token_access
