+++ addons/microsoft_calendar/tests/common.py
@@ -1,3 +1,5 @@
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 import pytz
 from datetime import datetime, timedelta
 from markupsafe import Markup
