+++ addons/microsoft_calendar/tests/test_delete_events.py
@@ -1,4 +1,5 @@
-# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from unittest.mock import patch, ANY, call
 from datetime import timedelta
