+++ addons/microsoft_calendar/tests/test_microsoft_event.py
@@ -1,3 +1,5 @@
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from datetime import datetime
 from dateutil.relativedelta import relativedelta
 from pytz import UTC
