+++ addons/microsoft_calendar/tests/test_microsoft_service.py
@@ -1,3 +1,5 @@
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 import json
 import requests
 from unittest.mock import patch, call, MagicMock
