+++ addons/microsoft_calendar/tests/test_update_events.py
@@ -1,4 +1,5 @@
-# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from datetime import datetime, timedelta
 from dateutil.parser import parse
 import logging
