+++ addons/microsoft_calendar/utils/__init__.py
@@ -1,4 +1,4 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from . import microsoft_calendar
 from . import microsoft_event
