+++ addons/microsoft_calendar/utils/microsoft_calendar.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 import requests
