+++ addons/microsoft_calendar/utils/microsoft_event.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from collections import abc
 from typing import Iterator, Mapping
